pkg/morph: make read network's magic `uint32`
When I wanted to use it, I wrote all the code based on `uint32` (because I knew
it should be this way), and at the last moment I found that `client` package
returns it as `uint64` for no reason.

+++ cmd/neofs-node/netmap.go
@@ -388,7 +388,7 @@ type netInfo struct {
 	netState netmap.State
 
 	magic interface {
-		MagicNumber() (uint64, error)
+		MagicNumber() (uint32, error)
 	}
 
 	morphClientNetMap *nmClient.Client
@@ -404,7 +404,7 @@ func (n *netInfo) Dump(ver version.Version) (*netmapSDK.NetworkInfo, error) {
 
 	var ni netmapSDK.NetworkInfo
 	ni.SetCurrentEpoch(n.netState.CurrentEpoch())
-	ni.SetMagicNumber(magic)
+	ni.SetMagicNumber(uint64(magic))
 
 	netInfoMorph, err := n.morphClientNetMap.ReadNetworkConfiguration()
 	if err != nil {
